% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transaction.R
\name{Transaction}
\alias{Transaction}
\title{Transaction() Function}
\usage{
Transaction(serviceKey, prod, trans, LAWD_CD, DEAL_YMD)
}
\arguments{
\item{serviceKey, }{prod, trans, LAWD_CD, DEAL_YMD}
}
\description{
국토교통부 부동산 실거래가 Open API
}
\examples{
Transaction("<service key>", "아파트", "매매", "41135", "202201")
}
\keyword{pdr}
